Big improvement
Improved error handler
Notifications added
Voice acting
Complete command handler, custom commands

+++ trayMenu.py
@@ -3,7 +3,10 @@ from PIL import Image
 from server import run_server, stop_server, open_browser
 from multiprocessing import Process
 from commandProcessor import run_command_processor
+from notificationModule import show_notification
 import time
+
+
 
 class TrayMenu:
     # init object
@@ -19,17 +22,20 @@ class TrayMenu:
         self.server_stopper = None
         self.command_process = None
 
+        show_notification("Голосовий помічник","Увімкніть прослуховування голосу")
+
+
     # start\stop action handler
     def start_stopVA(self, icon, item):
         self.changeState()
-        print(self.VAstate)
         if self.VAstate == "Start":
             self.command_process = Process(target=run_command_processor)
             self.command_process.start()
         else:
             self.command_process.terminate()
             self.command_process.join()
-        print('VA ' + self.VAstate + 'ed')
+            show_notification("Голосовий помічник","Ваш голос більше не слухається")
+
 
     #change state method
     def changeState(self):
@@ -37,6 +43,7 @@ class TrayMenu:
             self.VAstate = "Stop"
         else:
             self.VAstate = "Start"
+
 
     # open setting action handler
     def openSettings(self):
@@ -46,32 +53,31 @@ class TrayMenu:
             server_pid = self.server_process.pid
             self.server_stopper = Process(target=stop_server, kwargs = {"server_process":server_pid})
             self.server_stopper.start()
-            print("Запустил")
             return 1
+
         if self.server_process.is_alive() and self.server_stopper.is_alive():
             open_browser()
-            print("Уже существует")
             return 1
         else:
-            print("Запуск заново")
             self.server_process = None
             self.server_stopper = None
             self.openSettings()
 
 
-
-
     # quit action handler
     def on_quit(self, icon, item):
-        if self.server_process.is_alive():
-            self.server_process.terminate()
-            self.server_process.join()
-        if self.server_stopper.is_alive():
-            self.server_stopper.terminate()
-            self.server_stopper.join()
-        if self.command_process.is_alive():
-            self.command_process.terminate()
-            self.command_process.join()
+        if self.server_process != None:
+            if self.server_process.is_alive():
+                self.server_process.terminate()
+                self.server_process.join()
+        if self.server_stopper != None:
+            if self.server_stopper.is_alive():
+                self.server_stopper.terminate()
+                self.server_stopper.join()
+        if self.command_process != None:
+            if self.command_process.is_alive():
+                self.command_process.terminate()
+                self.command_process.join()
 
         icon.stop()
 
@@ -86,8 +92,3 @@ class TrayMenu:
 def startTray():
     tray_menu = TrayMenu()
     tray_menu.run()
-
-
-
-
-
